# api/openapi.yaml
openapi: 3.0.0
info:
  description: API for the Corp Fees central.
  title: CorpFees
  version: 1.0.0
servers:
- url: http://localhost:3000/
paths:
  /services:
    get:
      description: Get a list of all supported services
      operationId: getAllServices
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/paths/~1services/post/responses/201/content/application~1json/schema'
                type: array
          description: OK
    post:
      operationId: createService
      requestBody:
        $ref: '#/components/requestBodies/CreateServiceSchema'
        content:
          application/json:
            schema:
              example:
                title: title
              properties:
                title:
                  type: string
              required:
              - title
              title: CreateServiceSchema
              type: object
        description: This creates a service like Ammended and Restated articles, or
          Formation.
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponse'
          description: Created
  /services/{serviceId}/:
    delete:
      description: Delete a service.
      operationId: deleteService
      parameters:
      - explode: false
        in: path
        name: serviceId
        required: true
        schema:
          title: ServiceId
          type: string
        style: simple
      responses:
        "204":
          description: Deleted
    get:
      description: Get one service's details
      operationId: getService
      parameters:
      - explode: false
        in: path
        name: serviceId
        required: true
        schema:
          title: ServiceId
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/paths/~1services/post/responses/201/content/application~1json/schema'
          description: OK
    patch:
      description: Update Service values
      operationId: updateService
      parameters:
      - explode: false
        in: path
        name: serviceId
        required: true
        schema:
          title: ServiceId
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/UpdateServiceSchema'
        content:
          application/json:
            schema:
              example:
                title: title
              properties:
                title:
                  type: string
              title: UpdateServiceSchema
              type: object
        required: true
      responses:
        "204":
          description: Updated
  /attributes:
    get:
      description: Get all database wide attributes like stock count, jurisdiction
        and entity type.
      operationId: getAllAttributes
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AttributeResponse'
                type: array
          description: OK
    post:
      description: Create a database-wide attribute. Think stock count or entity type
      operationId: createAttribute
      requestBody:
        $ref: '#/components/requestBodies/CreateAttributeSchema'
        content:
          application/json:
            schema:
              example:
                title: title
              properties:
                title:
                  type: string
              required:
              - title
              title: CreateAttributeSchema
              type: object
        required: true
      responses:
        "201":
          description: Created
  /attributes/{attributeId}:
    delete:
      description: Delete an attribute.
      operationId: deleteAttribute
      parameters:
      - explode: false
        in: path
        name: attributeId
        required: true
        schema:
          title: AttributeId
          type: string
        style: simple
      responses:
        "204":
          description: Deleted
    patch:
      description: Update an attribute's name or other details. But not the attribute
        values.
      operationId: updateAttribute
      parameters:
      - explode: false
        in: path
        name: attributeId
        required: true
        schema:
          title: AttributeId
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/UpdateAttributeSchema'
        content:
          application/json:
            schema:
              example:
                title: title
              properties:
                title:
                  type: string
              title: UpdateAttributeSchema
              type: object
        required: true
      responses:
        "204":
          description: Updated
  /attributes/{attributeId}/values:
    description: create a new value and get all values assoc. w an attribute
    get:
      description: Get all a given attribute's possible values
      operationId: getAllAttributeValues
      parameters:
      - explode: false
        in: path
        name: attributeId
        required: true
        schema:
          title: AttributeId
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AttributeValueResponse'
                type: array
          description: OK
    post:
      description: Create a database-wide attribute value. Think 1000+ for stock count
        or Delware for jurisdiction.
      operationId: createAttributeValue
      parameters:
      - explode: false
        in: path
        name: attributeId
        required: true
        schema:
          title: AttributeId
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/CreateAttributeValueSchema'
        content:
          application/json:
            schema:
              example:
                title: title
              properties:
                title:
                  type: string
              required:
              - attributeId
              title: CreateAttributeValueSchema
              type: object
        required: true
      responses:
        "201":
          description: Created
  /attribute_values/{valueId}:
    delete:
      description: Delete an attribute value.
      operationId: deleteAttributeValue
      parameters:
      - explode: false
        in: path
        name: valueId
        required: true
        schema:
          title: ValueId
          type: string
        style: simple
      responses:
        "204":
          description: Deleted
    description: update and delete specific attribute values here
    patch:
      description: Update an attribute value.
      operationId: updateAttributeValue
      parameters:
      - explode: false
        in: path
        name: valueId
        required: true
        schema:
          title: ValueId
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/UpdateAttributeValueSchema'
        content:
          application/json:
            schema:
              example:
                title: title
              properties:
                title:
                  type: string
              title: UpdateAttributeValueSchema
              type: object
        required: true
      responses:
        "204":
          description: Updated
  /services/{serviceId}/attributes/{attributeId}/lines:
    get:
      operationId: getServiceAttribtuteLine
      parameters:
      - explode: false
        in: path
        name: serviceId
        required: true
        schema:
          title: ServiceId
          type: string
        style: simple
      - explode: false
        in: path
        name: attributeId
        required: true
        schema:
          title: AttributeId
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceAttributeLineResponse'
          description: Success
      summary: Given a service id and an attribute id, return the service attribute
        line for that combination
    post:
      operationId: createServiceAttributeLine
      parameters:
      - explode: false
        in: path
        name: serviceId
        required: true
        schema:
          title: ServiceId
          type: string
        style: simple
      - explode: false
        in: path
        name: attributeId
        required: true
        schema:
          title: AttributeId
          type: string
        style: simple
      responses:
        "201":
          description: Created
      summary: Create an attribute line for a service. This is how we relate attribute
        values to services.
  /service_attribute_lines/{lineId}:
    delete:
      operationId: deleteServiceAttributeLine
      parameters:
      - explode: false
        in: path
        name: lineId
        required: true
        schema:
          title: LineId
          type: string
        style: simple
      responses:
        "204":
          description: Deleted
      summary: Delete a service attribute line.
  /service_attribute_lines/{lineId}/values:
    description: This endpoint supports CRD on the allowed values for a service. Think
      1 day or 2 day or 3 day for filing speed attribute on Ammended and Restated
      articles.
    get:
      operationId: getServiceAttrVals
      parameters:
      - explode: false
        in: path
        name: lineId
        required: true
        schema:
          title: LineId
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ServiceAttributeValueResponse'
                type: array
          description: Created
      summary: Get all the service attribute values for a particular service attribute
        line.
    post:
      operationId: createServiceAttributeValue
      parameters:
      - explode: false
        in: path
        name: lineId
        required: true
        schema:
          title: LineId
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/CreateServiceAttributeValueSchema'
        content:
          application/json:
            schema:
              example:
                attributeValueId: attributeValueId
              properties:
                attributeValueId:
                  type: string
              required:
              - attributeValueId
              title: CreateServiceAttributeValueSchema
              type: object
        required: true
      responses:
        "201":
          description: Created
      summary: create a new service attribute value (not an attribute value.) This
        only applies to the service listed in the path.
  /services_attribute_values/{valueId}:
    delete:
      operationId: deleteServiceAttributeValue
      parameters:
      - explode: false
        in: path
        name: valueId
        required: true
        schema:
          title: ValueId
          type: string
        style: simple
      responses:
        "204":
          description: Deleted
      summary: Delete a service attribute value. valueId here is the service attribute
        value id NOT the attribute value id.
  /service_variants/:
    get:
      operationId: getVariants
      parameters:
      - description: This param is an array of strings where the strings are the service
          attribute value ids.
        explode: true
        in: query
        name: serviceAttributeValueIds[]
        required: false
        schema:
          items:
            type: string
          title: ServiceAttributeValueIds
          type: array
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VariantResponse'
          description: OK
      summary: Retrieve the variant (fee and other information) for a particular service
        variant, ie. (Amended and Restated Articles in Delaware, 1 Day). If you provide
        no values for the serviceAttribute Values then you will recieve a list of
        all service variants.
    post:
      operationId: createVariant
      requestBody:
        $ref: '#/components/requestBodies/CreateServiceVariantSchema'
        content:
          application/json:
            schema:
              example:
                serviceAttributeValueIds:
                - serviceAttributeValueIds
                - serviceAttributeValueIds
                fee: 0.8008281904610115
                serviceId: serviceId
              properties:
                serviceId:
                  type: string
                fee:
                  type: number
                serviceAttributeValueIds:
                  items:
                    type: string
                  type: array
              required:
              - fee
              - serviceAttributeValueIds
              - serviceId
              title: CreateServiceVariantSchema
              type: object
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/VariantCreatedResponse'
                type: array
          description: Created
      summary: Create a variant for a service with a specific combination of Jurisdiction,
        Filing Speed and Entity type.
  /service_variants/{variantId}:
    delete:
      operationId: deleteVariant
      parameters:
      - explode: false
        in: path
        name: variantId
        required: true
        schema:
          title: VariantId
          type: string
        style: simple
      responses:
        "204":
          description: Deleted
      summary: Delete a service variant.
...
components:
  requestBodies:
    UpdateAttributeValueSchema:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdateAttributeValueSchema'
      required: true
    CreateAttributeSchema:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateAttributeSchema'
      required: true
    UpdateAttributeSchema:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdateAttributeSchema'
      required: true
    CreateServiceSchema:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateServiceSchema'
      required: true
    CreateServiceAttributeValueSchema:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateServiceAttributeValueSchema'
      required: true
    CreateAttributeValueSchema:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateAttributeValueSchema'
      required: true
    UpdateServiceSchema:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdateServiceSchema'
      required: true
    CreateServiceVariantSchema:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateServiceVariantSchema'
      required: true
  schemas:
    ServiceAttributeLineResponse:
      example:
        id: id
      properties:
        id:
          type: string
      required:
      - id
      title: ServiceAttributeLineResponse
      type: object
...
    AttributeValueResponse:
      example:
        attributeId: attributeId
        id: id
        title: title
      properties:
        title:
          type: string
        id:
          type: string
        attributeId:
          type: string
      required:
      - attributeValueId
      - id
      - title
      title: AttributeValueResponse
      type: object
    VariantCreatedResponse:
      example:
        id: id
      properties:
        id:
          type: string
      required:
      - id
      title: VariantCreatedResponse
      type: object
    ServiceResponse:
      example:
        id: id
        title: title
      properties:
        id:
          type: string
        title:
          type: string
      required:
      - id
      - title
      title: ServiceResponse
      type: object
...
    ServiceAttributeValueResponse:
      example:
        value_title: value_title
        id: id
      properties:
        id:
          type: string
        value_title:
          type: string
      required:
      - id
      - value_title
      title: ServiceAttributeValueResponse
      type: object
    VariantResponse:
      example:
        fee: 0.8008281904610115
        id: id
        serviceId: serviceId
      properties:
        id:
          type: string
        fee:
          type: number
        serviceId:
          type: string
      required:
      - fee
      - id
      - serviceId
      title: VariantResponse
      type: object
    AttributeResponse:
      example:
        id: id
        title: title
      properties:
        id:
          type: string
        title:
          type: string
      required:
      - id
      - title
      title: AttributeResponse
      type: object
...